alice/paskills/discovery/dataset_generation/requirements.txt
certifi==2019.9.11
chardet==3.0.4
cyson==1.14.1
decorator==4.2.1
idna==2.8
networkx==1.11
nile==2.17.1
nile-debug==0.2.0
pycodestyle==2.5.0
Pygments==2.4.2
pyparsing==2.3.0
pytz==2019.1
PyYAML==5.1.2
qb2-core==1.21.0
requests==2.22.0
simplejson==3.17.0
six==1.13.0
statbox-bindings2==0.25.1
tabulate==0.7.5
urllib3==1.25.7
wcwidth==0.1.6
ws4py==0.3.4
yandex-yt==0.9.23
yql==1.2.91
